+++ bids/reports/parsing.py
@@ -5,214 +5,15 @@ Parsing functions for generating the MRI data acquisition portion of a
 methods section from a BIDS dataset.
 """
 import logging
-import os.path as op
 
-import math
+import nibabel as nib
 from num2words import num2words
 
 from .. import __version__
-from .utils import (num_to_str, list_to_str, get_slice_info,
-                    get_seqstr, get_sizestr)
+from . import parameters, utils
 
 logging.basicConfig()
 LOGGER = logging.getLogger('pybids.reports.parsing')
-
-
-def general_acquisition_info(metadata):
-    """
-    General sentence on data acquisition. Should be first sentence in MRI data
-    acquisition section.
-
-    Parameters
-    ----------
-    metadata : :obj:`dict`
-        The metadata for the dataset.
-
-    Returns
-    -------
-    out_str : :obj:`str`
-        Output string with scanner information.
-    """
-    out_str = ('MR data were acquired using a {tesla}-Tesla {manu} {model} '
-               'MRI scanner.')
-    out_str = out_str.format(tesla=metadata.get('MagneticFieldStrength',
-                                                'UNKNOWN'),
-                             manu=metadata.get('Manufacturer', 'MANUFACTURER'),
-                             model=metadata.get('ManufacturersModelName',
-                                                'MODEL'))
-    return out_str
-
-
-def get_slice_str(img, metadata):
-    if 'SliceTiming' in metadata.keys():
-        slice_order = ' in {0} order'.format(get_slice_info(metadata['SliceTiming']))
-        n_slices = len(metadata['SliceTiming'])
-    else:
-        slice_order = ''
-        n_slices = img.shape[2]
-    slice_str = '{n_slices} slices{slice_order}'.format(
-        n_slices=n_slices,
-        slice_order=slice_order
-    )
-    return slice_str
-
-
-def get_tr_str(metadata):
-    tr = metadata['RepetitionTime'] * 1000
-    tr = num_to_str(tr)
-    tr_str = 'repetition time, TR={tr}ms'.format(tr=tr)
-    return tr_str
-
-
-def get_duration(img, metadata):
-    tr = metadata['RepetitionTime']
-    n_tps = img.shape[3]
-    run_secs = math.ceil(n_tps * tr)
-    mins, secs = divmod(run_secs, 60)
-    duration = '{0}:{1:02.0f}'.format(int(mins), int(secs))
-    return duration
-
-
-def get_mbfactor_str(metadata):
-    if metadata.get('MultibandAccelerationFactor', 1) > 1:
-        mb_str = 'MB factor={}'.format(metadata['MultibandAccelerationFactor'])
-    else:
-        mb_str = ''
-    return mb_str
-
-
-def get_echotimes_str(metadata):
-    """Build a description of echo times from metadata field.
-
-    Parameters
-    ----------
-    metadata : dict
-        Metadata information for multiple files merged into one dictionary.
-        For multi-echo data, EchoTime should be a list.
-
-    Returns
-    -------
-    te_str : str
-        Description of echo times.
-    me_str : str
-        Whether the data are multi-echo or single-echo.
-    """
-    if 'EchoTime' in metadata.keys():
-        if isinstance(metadata['EchoTime'], list):
-            te = [num_to_str(t*1000) for t in metadata['EchoTime']]
-            te = list_to_str(te)
-            me_str = 'multi-echo'
-        else:
-            te = num_to_str(metadata['EchoTime']*1000)
-            me_str = 'single-echo'
-    else:
-        te = 'UNKNOWN'
-        me_str = 'UNKNOWN-echo'
-    te_str = 'echo time, TE={}ms'.format(te)
-    return te_str, me_str
-
-
-def get_size_strs(img):
-    """Build descriptions from sizes of imaging data, including field of view,
-    voxel size, and matrix size.
-
-    Parameters
-    ----------
-    img : nibabel.nifti1.Nifti1Image
-        Image object from which to determine sizes.
-
-    Returns
-    -------
-    fov_str
-    matrixsize_str
-    voxelsize_str
-    """
-    vs_str, ms_str, fov_str = get_sizestr(img)
-    fov_str = 'field of view, FOV={}mm'.format(fov_str)
-    voxelsize_str = 'voxel size={}mm'.format(vs_str)
-    matrixsize_str = 'matrix size={}'.format(ms_str)
-    return fov_str, matrixsize_str, voxelsize_str
-
-
-def get_inplaneaccel_str(metadata):
-    if metadata.get('ParallelReductionFactorInPlane', 1) > 1:
-        pr_str = ('in-plane acceleration factor='
-                  '{}'.format(metadata['ParallelReductionFactorInPlane']))
-    else:
-        pr_str = ''
-    return pr_str
-
-
-def get_flipangle_str(metadata):
-    return 'flip angle, FA={}<deg>'.format(metadata.get('FlipAngle', 'UNKNOWN'))
-
-
-def get_nvecs_str(img):
-    return '{} diffusion directions'.format(img.shape[3])
-
-
-def get_bval_str(bval_file):
-    # Parse bval file
-    with open(bval_file, 'r') as file_object:
-        d = file_object.read().splitlines()
-    bvals = [item for sublist in [l.split(' ') for l in d] for item in sublist]
-    bvals = sorted([int(v) for v in set(bvals)])
-    bvals = [num_to_str(v) for v in bvals]
-    bval_str = list_to_str(bvals)
-    bval_str = 'b-values of {} acquired'.format(bval_str)
-    return bval_str
-
-
-def get_dir_str(metadata, config):
-    dir_str = config['dir'][metadata['PhaseEncodingDirection']]
-    dir_str = 'phase encoding: {}'.format(dir_str)
-    return dir_str
-
-
-def get_for_str(metadata, layout):
-    if 'IntendedFor' in metadata.keys():
-        scans = metadata['IntendedFor']
-        run_dict = {}
-        for scan in scans:
-            fn = op.basename(scan)
-            if_file = [f for f in layout.get(extension=[".nii", ".nii.gz"]) if fn in f.path][0]
-            run_num = int(if_file.run)
-            target_type = if_file.entities['suffix'].upper()
-            if target_type == 'BOLD':
-                iff_meta = layout.get_metadata(if_file.path)
-                task = iff_meta.get('TaskName', if_file.entities['task'])
-                target_type_str = '{0} {1} scan'.format(task, target_type)
-            else:
-                target_type_str = '{0} scan'.format(target_type)
-
-            if target_type_str not in run_dict.keys():
-                run_dict[target_type_str] = []
-            run_dict[target_type_str].append(run_num)
-
-        for scan in run_dict.keys():
-            run_dict[scan] = [num2words(r, ordinal=True) for r in sorted(run_dict[scan])]
-
-        out_list = []
-        for scan in run_dict.keys():
-            if len(run_dict[scan]) > 1:
-                s = 's'
-            else:
-                s = ''
-            run_str = list_to_str(run_dict[scan])
-            string = '{rs} run{s} of the {sc}'.format(
-                rs=run_str, s=s, sc=scan)
-            out_list.append(string)
-        for_str = ' for the {0}'.format(list_to_str(out_list))
-    else:
-        for_str = ''
-    return for_str
-
-
-def clean_desc(desc):
-    desc = desc.replace('\n', ' ').lstrip()
-    while '  ' in desc:
-        desc = desc.replace('  ', ' ')
-    return desc
 
 
 def func_info(layout, files, config):
@@ -239,23 +40,29 @@ def func_info(layout, files, config):
         A description of the scan's acquisition information.
     """
     first_file = files[0]
+    metadata = first_file.get_metadata()
+    img = nib.load(first_file.path)
 
     # General info
-    task_name = metadata.get('TaskName', task+' task')
-    seqs, variants = get_seqstr(config, metadata)
+    task_name = first_file.get_entities()['task'] + ' task'
+    task_name = metadata.get('TaskName', task_name)
+    seqs, variants = parameters.get_seqstr(metadata, config)
+    all_runs = sorted(list(set([f.get_entities().get('run', 1) for f in files])))
+    n_runs = len(all_runs)
     if n_runs == 1:
         run_str = '{0} run'.format(num2words(n_runs).title())
     else:
         run_str = '{0} runs'.format(num2words(n_runs).title())
+    dur_str = parameters.get_dur_str(files)
 
     # Parameters
-    slice_str = get_slice_str(img, metadata)
-    tr_str = get_tr_str(metadata)
-    te_str, me_str = get_echotimes_str(metadata)
-    fa_str = get_flipangle_str(metadata)
-    fov_str, matrixsize_str, voxelsize_str = get_size_strs(img)
-    mb_str = get_mbfactor_str(metadata)
-    inplaneaccel_str = get_inplaneaccel_str(metadata)
+    slice_str = parameters.get_slice_str(img, metadata)
+    tr_str = parameters.get_tr_str(metadata)
+    te_str, me_str = parameters.get_echotimes_str(files)
+    fa_str = parameters.get_flipangle_str(metadata)
+    fov_str, matrixsize_str, voxelsize_str = parameters.get_size_strs(img)
+    mb_str = parameters.get_mbfactor_str(metadata)
+    inplaneaccel_str = parameters.get_inplaneaccel_str(metadata)
 
     parameters_str = [slice_str, tr_str, te_str, fa_str,
                       fov_str, matrixsize_str, voxelsize_str,
@@ -265,23 +72,20 @@ def func_info(layout, files, config):
 
     desc = '''
            {run_str} of {task} {variants} {seqs} {me_str} fMRI data were
-           collected ({parameters_str}).
-           Run duration was {duration} minutes, during which
-           {n_vols} functional volumes were acquired.
+           collected ({parameters_str}). {dur_str}
            '''.format(run_str=run_str,
                       task=task_name,
                       variants=variants,
                       seqs=seqs,
                       me_str=me_str,
                       parameters_str=parameters_str,
-                      duration=get_duration(img, metadata),
-                      n_vols=img.shape[3],
+                      dur_str=dur_str,
                       )
-    desc = clean_desc(desc)
+    desc = utils.clean_multiline(desc)
     return desc
 
 
-def anat_info(suffix, metadata, img, config):
+def anat_info(layout, files, config):
     """
     Generate a paragraph describing T1- and T2-weighted structural scans.
 
@@ -303,15 +107,26 @@ def anat_info(suffix, metadata, img, config):
     desc : :obj:`str`
         A description of the scan's acquisition information.
     """
+    first_file = files[0]
+    metadata = first_file.get_metadata()
+    img = nib.load(first_file.path)
+
     # General info
-    seqs, variants = get_seqstr(config, metadata)
+    seqs, variants = parameters.get_seqstr(metadata, config)
+    all_runs = sorted(list(set([f.get_entities().get('run', 1) for f in files])))
+    n_runs = len(all_runs)
+    if n_runs == 1:
+        run_str = '{0} run'.format(num2words(n_runs).title())
+    else:
+        run_str = '{0} runs'.format(num2words(n_runs).title())
+    scan_type = first_file.get_entities()['suffix'].replace('w', '-weighted')
 
     # Parameters
-    slice_str = get_slice_str(img, metadata)
-    tr_str = get_tr_str(metadata)
-    te_str, me_str = get_echotimes_str(metadata)
-    fa_str = get_flipangle_str(metadata)
-    fov_str, matrixsize_str, voxelsize_str = get_size_strs(img)
+    slice_str = parameters.get_slice_str(img, metadata)
+    tr_str = parameters.get_tr_str(metadata)
+    te_str, me_str = parameters.get_echotimes_str(files)
+    fa_str = parameters.get_flipangle_str(metadata)
+    fov_str, matrixsize_str, voxelsize_str = parameters.get_size_strs(img)
 
     parameters_str = [slice_str, tr_str, te_str, fa_str,
                       fov_str, matrixsize_str, voxelsize_str]
@@ -319,20 +134,20 @@ def anat_info(suffix, metadata, img, config):
     parameters_str = '; '.join(parameters_str)
 
     desc = '''
-           {n_scans} {suffix} {variants} {seqs} {me_str} structural MRI scan(s)
-           were collected ({parameters_str}).
-           '''.format(n_scans='One',
-                      suffix=suffix,
+           {run_str} of {scan_type} {variants} {seqs} {me_str} structural MRI
+           data were collected ({parameters_str}).
+           '''.format(run_str=run_str,
+                      scan_type=scan_type,
                       variants=variants,
                       seqs=seqs,
                       me_str=me_str,
                       parameters_str=parameters_str
                       )
-    desc = clean_desc(desc)
+    desc = utils.clean_multiline(desc)
     return desc
 
 
-def dwi_info(bval_file, metadata, img, config):
+def dwi_info(layout, files, config):
     """
     Generate a paragraph describing DWI scan acquisition information.
 
@@ -354,19 +169,30 @@ def dwi_info(bval_file, metadata, img, config):
     desc : :obj:`str`
         A description of the DWI scan's acquisition information.
     """
+    first_file = files[0]
+    metadata = first_file.get_metadata()
+    img = nib.load(first_file.path)
+    bval_file = first_file.replace('.nii.gz', '.bval').replace('.nii', '.bval')
+
     # General info
-    seqs, variants = get_seqstr(config, metadata)
+    seqs, variants = parameters.get_seqstr(metadata, config)
+    all_runs = sorted(list(set([f.get_entities().get('run', 1) for f in files])))
+    n_runs = len(all_runs)
+    if n_runs == 1:
+        run_str = '{0} run'.format(num2words(n_runs).title())
+    else:
+        run_str = '{0} runs'.format(num2words(n_runs).title())
 
     # Parameters
-    tr_str = get_tr_str(metadata)
-    te_str, me_str = get_echotimes_str(metadata)
-    fa_str = get_flipangle_str(metadata)
-    fov_str, voxelsize_str, matrixsize_str = get_size_strs(img)
-    bval_str = get_bval_str(bval_file)
-    nvec_str = get_nvecs_str(img)
-    mb_str = get_mbfactor_str(metadata)
+    tr_str = parameters.get_tr_str(metadata)
+    te_str, me_str = parameters.get_echotimes_str(files)
+    fa_str = parameters.get_flipangle_str(metadata)
+    fov_str, voxelsize_str, matrixsize_str = parameters.get_size_strs(img)
+    bval_str = parameters.get_bval_str(bval_file)
+    nvec_str = parameters.get_nvecs_str(img)
+    mb_str = parameters.get_mbfactor_str(metadata)
 
-    parameters_str = [slice_str, tr_str, te_str, fa_str, fov_str,
+    parameters_str = [tr_str, te_str, fa_str, fov_str,
                       matrixsize_str, voxelsize_str,
                       bval_str, nvec_str,
                       mb_str]
@@ -374,17 +200,18 @@ def dwi_info(bval_file, metadata, img, config):
     parameters_str = '; '.join(parameters_str)
 
     desc = '''
-           One run of {variants} {seqs} diffusion-weighted (dMRI) data were
+           {run_str} of {variants} {seqs} diffusion-weighted (dMRI) data were
            collected ({parameters_str}).
-           '''.format(variants=variants,
+           '''.format(run_str=run_str,
+                      variants=variants,
                       seqs=seqs,
                       parameters_str=parameters_str
                       )
-    desc = clean_desc(desc)
+    desc = utils.clean_multiline(desc)
     return desc
 
 
-def fmap_info(metadata, img, config, layout):
+def fmap_info(layout, files, config):
     """
     Generate a paragraph describing field map acquisition information.
 
@@ -404,17 +231,27 @@ def fmap_info(metadata, img, config, layout):
     desc : :obj:`str`
         A description of the field map's acquisition information.
     """
+    first_file = files[0]
+    metadata = first_file.get_metadata()
+    img = nib.load(first_file.path)
+
     # General info
-    seqs, variants = get_seqstr(config, metadata)
+    seqs, variants = parameters.get_seqstr(metadata, config)
+    all_runs = sorted(list(set([f.get_entities().get('run', 1) for f in files])))
+    n_runs = len(all_runs)
+    if n_runs == 1:
+        run_str = '{0} run'.format(num2words(n_runs).title())
+    else:
+        run_str = '{0} runs'.format(num2words(n_runs).title())
 
     # Parameters
-    dir_str = get_dir_str(metadata, config)
-    slice_str = get_slice_str(img, metadata)
-    tr_str = get_tr_str(metadata)
-    te_str, me_str = get_echotimes_str(metadata)
-    fa_str = get_flipangle_str(metadata)
-    fov_str, matrixsize_str, voxelsize_str = get_size_strs(img)
-    mb_str = get_mbfactor_str(metadata)
+    dir_str = parameters.get_dir_str(metadata, config)
+    slice_str = parameters.get_slice_str(img, metadata)
+    tr_str = parameters.get_tr_str(metadata)
+    te_str, me_str = parameters.get_echotimes_str(files)
+    fa_str = parameters.get_flipangle_str(metadata)
+    fov_str, matrixsize_str, voxelsize_str = parameters.get_size_strs(img)
+    mb_str = parameters.get_mbfactor_str(metadata)
 
     parameters_str = [dir_str, slice_str, tr_str, te_str, fa_str,
                       fov_str, matrixsize_str, voxelsize_str,
@@ -422,7 +259,7 @@ def fmap_info(metadata, img, config, layout):
     parameters_str = [d for d in parameters_str if len(d)]
     parameters_str = '; '.join(parameters_str)
 
-    for_str = get_for_str(metadata, layout)
+    for_str = parameters.get_for_str(metadata, layout)
 
     desc = '''
            A {variants} {seqs} field map ({parameters_str}) was
@@ -432,8 +269,33 @@ def fmap_info(metadata, img, config, layout):
                       for_str=for_str,
                       parameters_str=parameters_str
                       )
-    desc = clean_desc(desc)
+    desc = utils.clean_multiline(desc)
     return desc
+
+
+def general_acquisition_info(metadata):
+    """
+    General sentence on data acquisition. Should be first sentence in MRI data
+    acquisition section.
+
+    Parameters
+    ----------
+    metadata : :obj:`dict`
+        The metadata for the dataset.
+
+    Returns
+    -------
+    out_str : :obj:`str`
+        Output string with scanner information.
+    """
+    out_str = ('MR data were acquired using a {tesla}-Tesla {manu} {model} '
+               'MRI scanner.')
+    out_str = out_str.format(tesla=metadata.get('MagneticFieldStrength',
+                                                'UNKNOWN'),
+                             manu=metadata.get('Manufacturer', 'MANUFACTURER'),
+                             model=metadata.get('ManufacturersModelName',
+                                                'MODEL'))
+    return out_str
 
 
 def final_paragraph(metadata):
@@ -462,49 +324,8 @@ def final_paragraph(metadata):
            automatically using pybids ({meth_vers}).
            '''.format(software_str=software_str,
                       meth_vers=__version__)
-    desc = clean_desc(desc)
+    desc = utils.clean_multiline(desc)
     return desc
-
-
-def collect_associated_files(layout, files, extra_entities=()):
-    """Collect and group BIDSFiles with multiple files per acquisition.
-
-    Parameters
-    ----------
-    layout
-    files : list of BIDSFile
-    extra_entities
-
-    Returns
-    -------
-    collected_files : list of list of BIDSFile
-    """
-    MULTICONTRAST_ENTITIES = ['echo', 'part', 'ch', 'direction']
-    MULTICONTRAST_SUFFICES = [
-        ('bold', 'phase'),
-        ('phase1', 'phase2', 'phasediff', 'magnitude1', 'magnitude2'),
-    ]
-    if len(extra_entities):
-        MULTICONTRAST_ENTITIES += extra_entities
-
-    collected_files = []
-    for f in files:
-        if len(collected_files) and any(f in filegroup for filegroup in collected_files):
-            continue
-        ents = f.get_entities()
-        ents = {k: v for k, v in ents.items() if k not in MULTICONTRAST_ENTITIES}
-
-        # Group files with differing multi-contrast entity values, but same
-        # everything else.
-        all_suffices = ents['suffix']
-        for mcs in MULTICONTRAST_SUFFICES:
-            if ents['suffix'] in mcs:
-                all_suffices = mcs
-                break
-        ents.pop('suffix')
-        associated_files = layout.get(suffix=all_suffices, **ents)
-        collected_files.append(associated_files)
-    return collected_files
 
 
 def parse_files(layout, data_files, sub, config, **kwargs):
@@ -526,7 +347,7 @@ def parse_files(layout, data_files, sub, config, **kwargs):
     kwargs = {k: v for k, v in kwargs.items() if v is not None}
 
     # Group files into individual runs
-    data_files = collect_associated_files(data_files, extra_entities=['run'])
+    data_files = utils.collect_associated_files(layout, data_files, extra_entities=['run'])
 
     description_list = []
     # Assume all data have same basic info
@@ -535,7 +356,7 @@ def parse_files(layout, data_files, sub, config, **kwargs):
         if group[0].entities['datatype'] == 'func':
             group_description = func_info(layout, group, config)
         elif (group[0].entities['datatype'] == 'anat') and \
-             data_file.entities['suffix'].endswith('w'):
+                group[0].entities['suffix'].endswith('w'):
             group_description = anat_info(layout, group, config)
         elif group[0].entities['datatype'] == 'dwi':
             group_description = dwi_info(layout, group, config)
